packages/backend/pyproject.toml
[tool.poetry]
name = "brace-backend"
version = "0.2.0"
description = "BRACE backend service for Telegram Mini App"
authors = ["Codex <codex@example.com>"]
readme = "README.md"
packages = [{ include = "brace_backend", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
SQLAlchemy = "^2.0.32"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
passlib = "^1.7.4"
loguru = "^0.7.2"
slowapi = "^0.1.9"
itsdangerous = "^2.2.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
ruff = "^0.5.4"
black = "^24.4.2"
factory-boy = "^3.3.0"
pytest-factoryboy = "^2.7.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
addopts = "--capture=tee-sys"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true  # <— переместили сюда!

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]
ignore = [
    "B008",  # FastAPI: использование Depends() в параметрах функций допустимо
    "E501",  # разрешаем длинные строки (до 100 символов, как в black)
    "F821",  # forward references в SQLAlchemy (User, Product и т.д.)
]
